# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TrueShot API",
    "description": "Intelligent 3D scanning and photogrammetry API",
    "license": {
      "name": ""
    },
    "version": "6.8.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/devices": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "Get all devices from all sources",
        "operationId": "get_all_devices",
        "responses": {
          "200": {
            "description": "Device list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnifiedDevice"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/devices/batch": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Batch action on all devices of a type",
        "operationId": "batch_device_action",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch action result",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/devices/stats": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "Get device statistics summary",
        "operationId": "get_device_stats",
        "responses": {
          "200": {
            "description": "Device stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceStats"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/devices/{id}/action": {
      "post": {
        "tags": [
          "devices"
        ],
        "operationId": "device_action",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Device id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Action result",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/phones/capture-all": {
      "post": {
        "tags": [
          "phones"
        ],
        "summary": "Trigger capture on ALL ready phones",
        "operationId": "capture_all_phones",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Capture broadcast",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}